Refactor home and add contact screens
